+++ smarts/core/agent_interface.py
@@ -268,12 +268,10 @@ class AgentInterface:
     debug: bool = False
     """Enable debug information for the various sensors and action spaces."""
 
-    event_configuration: EventConfiguration = field(
-        default_factory=lambda: EventConfiguration()
-    )
+    event_configuration: EventConfiguration = field(default_factory=EventConfiguration)
     """Configurable criteria of when to trigger events"""
 
-    done_criteria: DoneCriteria = field(default_factory=lambda: DoneCriteria())
+    done_criteria: DoneCriteria = field(default_factory=DoneCriteria)
     """Configurable criteria of when to mark this actor as done. Done actors will be
     removed from the environment and may trigger the episode to be done."""